# src/main/resources/api.yaml
swagger: '2.0'
info:
  description: User API
  version: 1.0.0
  title: User API
  contact:
    name: User service
tags:
  - name: User Service
    description: User Service
schemes:
  - https
paths:
  '/userService/users/':
      get:
          tags:
            - User
          summary: Get all users
          description: Get all users
          operationId: getAvailableUsers
          produces:
            - application/json
          responses:
            '200':
              description: Successful operation
              schema:
                $ref: '#/definitions/Users'
            '404':
              description: Incorrect data input
      post:
        tags:
          - User
        summary: Post given user name and password
        description: Create a new user
        operationId: postUser
        produces:
          - application/json
        parameters:
          - name: body
            in: body
            schema:
              $ref: "#/definitions/PostUserAccount"
        responses:
          '200':
            description: Successful operation
            schema:
              $ref: '#/definitions/UserModel'
          '404':
            description: Incorrect data input

  '/userService/users/{userName}':
    get:
      tags:
        - User
      summary: Validate user
      description: Validate user
      operationId: validateUser
      produces:
        - application/json
      parameters:
        - name: userName
          in: path
          type: string
          description: User name
          minLength: 1
          maxLength: 30
          required: true
        - name: password
          in: query
          description: User password
          type: string
          format: "password"
          minLength: 6
          maxLength: 12
          required: true
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/UserModel'
        '404':
          description: Incorrect user data input

definitions:
  UserModel:
    title: User Account data.
    description: User Account data.
    type: object
    properties:
      userId:
        type: integer
        format: int64
        title: User id
        description: User id for a user account.
        minimum: 1
        maximum: 9999999999
      userName:
        type: string
        title: User name
        description: User name
        minLength: 1
        maxLength: 30
      userAddress:
        type: string
        title: User address
        description: User address
        minLength: 1
        maxLength: 50
      password:
        type: string
        title: User password
        description: User password
        format: "password"
        minLength: 6
        maxLength: 12
      email:
        type: string
        title: Email address
        description: Email address
        minLength: 1
        maxLength: 50

    required:
      - userId

  PostUserAccount:
    title: Post user Account data.
    description: Post user Account data.
    type: object
    properties:
      userName:
        type: string
        title: User name
        description: User name
        minLength: 1
        maxLength: 30
      password:
        type: string
        title: User password
        description: User password
        format: "password"
        minLength: 6
        maxLength: 12
      userAddress:
        type: string
        title: User address
        description: User address
        minLength: 1
        maxLength: 50
      email:
        type: string
        title: Email address
        description: Email address
        minLength: 1
        maxLength: 50

    required:
      - userName
      - password

  Users:
    title: List of All Available Users
    description: List of All Available Users
    type: object
    properties:
      users:
        title: List of All Available Users
        description: List of All Available Users
        type: array
        items:
          $ref: '#/definitions/UserModel'